backend/requirements.txt
# Web
fastapi==0.109.0
uvicorn[standard]==0.27.0

# AI / LLM
openai==1.12.0
httpx==0.27.2
langchain==0.1.6
langchain-openai==0.0.5
langchain-community==0.0.20
langchain-core==0.1.23

# Document processing
pypdf==4.0.1
python-docx==1.1.0

# Vector store & embeddings
chromadb==0.4.22
sentence-transformers==2.2.2

# Face detection
opencv-python-headless==4.9.0.80
numpy==1.26.4

# Utils
python-multipart==0.0.6
python-dotenv==1.0.1
